refactored, fixed GH actions workflow tests
