+++ src/middleware/logger.js
@@ -2,6 +2,7 @@
 
 const logger = (req, res, next) => {
   console.log(req.method, req.url);
+  console.log('logger middleware');
   next();
 };
