+++ src/middleware/validator.js
@@ -1,9 +1,21 @@
 'use strict';
 
+// const validator = (req, res, next) => {
+//   console.log("validator");
+// };
+
+// module.export = {
+//   validator,
+// };
+
+
 const validator = (req, res, next) => {
-  console.log("validator");
+    if(!req.query.name){
+        next('This is the validator middleware: please enter a name in the query');
+    }
+    next();
 };
 
-module.export = {
-  validator,
-};
+module.exports = {
+    validator,
+}
