+++ src/server.js
@@ -1,14 +1,20 @@
 'use strict';
 
+// IMPORTS
+
 const express = require('express');
-const logger = require('./middleware/logger.js');
-const validator = require('./middleware/validator.js');
+const { logger } = require('./middleware/logger.js');
+const { validator } = require('./middleware/validator.js');
 
-const hello = require('./handler/hello.js');
-const data  = require('.handler/data.js');
+const { error } = require('./error-handlers/404.js');
+const { serverError } = require('./error-handlers/500.js');
 
-const error = require('./error-handlers/404.js');
-const serverError = require('./error-handlers/500.js');
+const app = express();
+
+app.use(express.json());
+// app.use(logger);
+
+// HANDLERS
 
 const hello = (req, res) => {
   res.status(200).send('Hello, World');
@@ -22,18 +28,35 @@ const data = (req, res) => {
 };
 
 const person = (req, res ) => {
-    res.status(200).send({name: req.params.name});
+    if(req.query.name){
+        res.status(200).send({name: req.params.name});
+    }
+    res.status(500).send('please enter a name');
 };
 
-const app = express();
 
-app.get('/', hello);
+app.get('/',  hello);
+// app.get('/', (req, res) => {
+//     res.status(200).send('Hello, World');
+// });
+
 app.get('/data', data);
+// app.get('/', (req, res) =>{
+//     res.status(200).send({
+//         name: 'Tony',
+//         role: 'Student',
+//     });
+// });
 
-app.use(logger);
-app.use(validator);
-app.use(error);
-app.use(serverError);
+// app.get('/person', validator, person);
+
+// app.use(logger);
+// app.use(validator);
+// app.use('*', error);
+// app.use(serverError);
+
+
+// FUNCTIONALITY TO GET START SERVER
 
 function start(port) {
   app.listen(port, () => console.log(`Server listening on port ${port}`));
